requirements.txt
blinker==1.8.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.3
pillow==10.4.0
pycparser==2.22
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
